requirements.txt
beautifulsoup4==4.5.1
chardet==2.3.0
Django==1.8.13
django-activeurl==0.1.9
django-appconf==1.0.2
django-ckeditor==5.1.1
django-classy-tags==0.8.0
django-debug-toolbar==1.4
django-extensions==1.7.4
django-haystack==2.5.0
django-ipware==1.1.5
django-jinja==2.2.1
django-markdown==0.8.4
django-mptt==0.8.6
django-nested-admin==3.0.10
django-pagedown==0.1.1
django-permanent==1.1.4
django-pure-pagination==0.3.0
django-resized==0.3.5
django-stdimage==2.3.3
django-stronghold==0.2.8
django-suit==0.2.21
django-webtest==1.8.0
django-widget-tweaks==1.4.1
Jinja2==2.8
ldif3==3.1.1
lxml==3.6.4
Markdown==2.6.7
MarkupSafe==0.23
mysqlclient==1.3.7
Pillow==3.3.1
progressbar2==3.10.1
python-card-me==0.9.3
python-dateutil==2.5.3
python-monkey-business==1.0.0
python-utils==2.0.0
six==1.10.0
sorl-thumbnail==12.4a1
sqlparse==0.2.1
surlex==0.2.0
unicodecsv==0.14.1
virtualenv==15.0.3
virtualenvwrapper-win==1.2.1
unidecode==0.4.19
waitress==1.0.0
WebOb==1.6.1
WebTest==2.0.23
Whoosh==2.7.4
XlsxWriter==0.9.3
django-sortedm2m==1.3.3
ExifRead==2.1.2